Add DerValue decoder class

+++ src/main/java/by/bcrypto/bee2j/DerAnchor.java
@@ -7,7 +7,7 @@ import com.sun.jna.Structure;
 @Structure.FieldOrder({"der", "pos", "tag", "len"})
 public class DerAnchor extends Structure implements Structure.ByReference {
 	public Pointer der;	    /*!< код */
-	public int pos;			/*!< позиция */
+	public long pos;			/*!< позиция */
 	public int tag;			/*!< тег */
-	public int len;			/*!< длина */
+	public long len;			/*!< длина */
 }
